SLING-13253 - address review: extract PomParser, assert bulk actions
- Extract POM parsing and reactor aggregation from RepositoryService into a new
  PomParser class, tested directly in PomParserTest (no reflection). Addresses
  the review note that this functionality warrants its own class.
- MockNexus now records the last staging bulk action, so testPromote/
  testCloseWithDescription/testDrop assert the command maps to the correct
  Nexus action (promote/close/delete) instead of verifying nothing.
